#69 Classを間違えている箇所があったので修正と、TestCase, WasRun　クラスをテストパッケージから除いた
