--- shino/src/test/java/org/codepolaris/tdd/chapter2/TestCase.java
@@ -1,9 +1,0 @@
-package org.codepolaris.tdd.chapter2;
-
-public class TestCase {
-  protected String testName;
-
-  public TestCase(String testName) {
-    this.testName = testName;
-  }
-}
