+++ shino/src/test/java/org/codepolaris/tdd/chapter2/TestCaseTest.java
@@ -5,13 +5,13 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class XunitTest {
+class TestCaseTest {
 
   @Test
   void testRunning() {
     val test = new WasRun("testMethod");
-    assertThat(test.getWasRun()).isNull();
+    assertThat(test.getHasRun()).isFalse();
     test.run();
-    assertThat(test.getWasRun()).isEqualTo(1);
+    assertThat(test.getHasRun()).isTrue();
   }
 }
